fix jenkinsfile

+++ Jenkinsfile
@@ -15,7 +15,26 @@ pipeline {
 				sh '''
 					docker-compose up --build -d db
 					echo "Attente de la base de données..."
-					sleep 15
+					
+					# Attend jusqu'à 2 minutes que la BDD soit prête
+					max_wait=120
+					count=0
+					while [ "$count" -lt "$max_wait" ]; do
+						status=$(docker inspect --format '{{.State.Health.Status}}' mysql_db)
+						if [ "$status" == "healthy" ]; then
+							echo "✅ Base de données prête."
+							break
+						fi
+						echo "En attente de la BDD (statut: $status)..."
+						sleep 5
+						count=$((count + 5))
+					done
+
+					if [ "$status" != "healthy" ]; then
+						echo "❌ La base de données n'a pas démarré à temps."
+						exit 1
+					fi
+
 					docker-compose run --rm app pytest
 				'''
 			}
